Mitigate race condition between McpsReq flag test/clear and flag set in another task
Right after testing the McpsReq flag and calling the confirmed-tx
callback, there is a race condition in which another task may find that the
confirmed-tx was terminated and attempts to schedule a brand-new
confirmed-tx, even if the callback itself does not schedule a new
transmission. After this, the code clears the just-checked McpsReq
flag, oblivious to the fact that the recently-scheduled confirmed
transmission relies on the flag being set. This eventually causes a
failure to call the confirmed-tx callback for the next finished
confirmed transmission.

This patch mitigates the race by copying the flags variable to a local
copy, then immediately clearing the global version, and only then test
the local copy. This greatly reduces the race window in which the task
might be preempted and a new transmission attempted.